Refactor Docker setup for interactive use and streamline volume bindings
This update modifies the Docker configuration to allow for interactive use by removing the command override in the Dockerfile and setting the entrypoint to bash. Additionally, the volume bindings in the docker-compose.yaml file have been updated to use absolute paths for music and playlists, enhancing clarity and usability. The main.py file has been removed as it is no longer needed.